+++ NEWS.md
@@ -15,6 +15,8 @@
 - Added a warning for potential infinite loops in processes without
   sensitivity and lacking any wait statements (from @NikLeberg).
 - Fixed a crash when `'last_value` is used with record types (#1043).
+- Selected names can now be to register value change callbacks with
+  `vhpi_register_cb` (#1045).
 - Several other minor bugs were resolved (#1038).
 
 ## Version 1.14.1 - 2024-10-26